use fmax/fmin here instead of fmaxf/fminf since we use doubles now

+++ sources/bvh/aabb.c
@@ -27,8 +27,8 @@ double	aabb_raycast(t_ray ray, t_bounds bounds)
 	double t5 = (bounds.min[2] - ray.O.z) * ray.rd.z;
 	double t6 = (bounds.max[2] - ray.O.z) * ray.rd.z;
 
-	double tmin = fmax(fmaxf(fmin(t1, t2), fmin(t3, t4)), fmin(t5, t6));
-	double tmax = fmin(fminf(fmax(t1, t2), fmax(t3, t4)), fmax(t5, t6));
+	double tmin = fmax(fmax(fmin(t1, t2), fmin(t3, t4)), fmin(t5, t6));
+	double tmax = fmin(fmin(fmax(t1, t2), fmax(t3, t4)), fmax(t5, t6));
 
 	if (tmax < 0 || tmin > tmax)
 		return (DBL_MAX);
